split cli and scanner
