+++ test/test_scanner.py
@@ -14,7 +14,7 @@ from sys import platform
 import pytest
 
 from cve_bin_tool.cvedb import CVEDB
-from cve_bin_tool.cli import Scanner, InvalidFileError
+from cve_bin_tool.scanner import Scanner, InvalidFileError
 from test.utils import download_file, LONG_TESTS
 from collections import defaultdict
 
@@ -64,18 +64,6 @@ class TestScanner:
         """ Run the scanner on a file """
         return self.scanner.scan_file(os.path.join(BINARIES_PATH, filename))
 
-    def test_vendor_package_pairs(self):
-        """ Test the vendor/package pairs for NVD lookup """
-        pairs = list(
-            self.scanner.vendor_package_pairs(
-                "\n".join(["VPkg: haxx, curl", "VPkg: feed, face"])
-            )
-        )
-        assert pairs
-        assert "haxx" in dict(pairs)
-        assert "curl" in dict(map(lambda x: x[::-1], pairs))
-        assert "feed" in dict(pairs)
-        assert "face" in dict(map(lambda x: x[::-1], pairs))
 
     def test_does_not_scan_symlinks(self):
         """ Test that the scanner doesn't scan symlinks """
@@ -118,13 +106,13 @@ class TestScanner:
         tempfile = os.path.join(self.tempdir, filename)
         download_file(url + filename, tempfile)
         # new scanner for the new test.
-        self.scanner = Scanner(cvedb=self.cvedb)
+        self.scanner = Scanner(cvedb=self.cvedb, should_extract=True)
         # run the tests
-        cves = self.scanner.extract_and_scan(tempfile)
+        self.scanner.recursive_scan(tempfile)
 
         # make sure we found the expected package/version
-        assert package in cves
-        assert version in cves[package]
+        assert package in self.scanner.all_cves
+        assert version in self.scanner.all_cves[package]
 
     @pytest.mark.parametrize(
         "binary, package, version, are_in, not_in",
